Fix section references
This fixes all section references except for a few referring to removed
chapters.

We can't use reST implicit section hyperlinks because those don't work
across files, so we have to use the sphinx style which requires manually
adding section labels.

The appendices are using the same section numbering instead of A, B, C,
etc. That doesn't appear to be supported by the toolchain AFAICT.

Signed-off-by: Rob Herring <robh@kernel.org>