Workaround overflowing table cell width problem
In some of the table cells there are literal values that are longer than
the width of the cell. Ideally, Latex would break the words into
multiple lines. It will do that for hyphenated words, but only in normal
text. It won't do it for literal blocks. Until this problem is solved,
work around the overflow issue by forcing the content of adjacent cells
to be moved down to the line below by using line blocks.

Signed-off-by: Grant Likely <grant.likely@linaro.org>