Set minimum Sphinx version to 1.2.3
v1.2.3 is packaged up to version 1.2.3 in Ubuntu 15.10, so it is at
least a useful baseline. Set that to the minimum version required to
build the document.

Although in reality 1.4 would be better. There are figure referencing
errors with 1.2 that aren't present in 1.4. However, that will make it
harder for some to build the document, so stick with 1.2.3 for now.

Signed-off-by: Grant Likely <grant.likely@linaro.org>