Rework revision history and acknowledgements
DTSpec is a new document. We don't need to keep the old acknowledgements
section from the ePAPR. Take list of names and companies out of the spec
document entirely. Instead, we will collect a list of contributors
whenever we do a new release announcement.

Also, take away the old ePAPR revision history. It isn't appropriate to
claim to be an immediate successor when there are portions of ePAPR
which don't carry over to this document.

Signed-off-by: Grant Likely <grant.likely@linaro.org>
Acked-by: Mark Rutland <mark.rutland@arm.com>
Acked-by: Stuart Yoder <stuart.yoder@nxp.com>
Cc: Stewart Smith <stewart@linux.vnet.ibm.com>