Add target to generate changebar pdfs
Add the "latexdiff" make target to generate a diff between the current
build and the previous release of the document. Type "make latexdiff" to
use this new target.

This patch is inelegant since it hard codes the file name, but it works.
It can be cleaned up to be more generic, but that isn't immediately
necessary.

Signed-off-by: Grant Likely <grant.likely@linaro.org>