Extract document version from git tag
Instead of hard coding the document version in the config file, use
'git describe' to get the version number from the tags and use that.
Just like before, if the version string contains a '-', then it is not
an official release and a watermark will be added to the PDF. This
catches prereleases "-pre*", untagged revisions "-g<hash>", and
uncommitted changes "-dirty".

Signed-off-by: Grant Likely <grant.likely@arm.com>
Reviewed-by: Rob Herring <robh@kernel.org>