Remove old build artifacts from repo
The releases/ subdirectory contains latex build artifacts. The idea was
to use it when creating a changebar version of the document, but it is a
little difficult to manage as part of the repo. The old document is
always available in a previous git tag anyway, so the build products can
be regenerated. So, drop the releases directory entirely.

Signed-off-by: Grant Likely <grant.likely@arm.com>