Clarify DTB section preable
The DTB chapter used to be titled the "Flattened Devicetree Structure"
chapter, but that seems to be unclear when it is talking about a
specific encoding of devicetree. In this case, the encoding is 'dtb'.
The source format chapter specifies "DTS", so change the start of this
chapter to line up with the DTS chapter.

At the same time, update the introduction to include the source format
chapter.

Signed-off-by: Grant Likely <grant.likely@arm.com>