Fix python3 compatibility and switch travis to python3
Python2 is EOL. Switch to py3. The move to python3 requires the 'typing'
package to be installed. Add it to the travisci file. Also update to
'focal' Ubuntu release to get a python newer than 3.3.

This patch reworks the use of git-describe to work out the document
version number. The existing code falls over when using Sphinx on Python
3 due to the string datatype changes. Wrap the output of subprocess with
str() to fix.

Finally, the changebar version build is currently failing and causing
latex to stall. Drop it for now until it can be resolved.

Signed-off-by: Grant Likely <grant.likely@arm.com>