Updates to match Sphinx templates
This should be a no-op commit. It restructures the config file to match
the style and format of the generic config file generated by
sphinx-quickstart. The only reason to do this is to make it easy to
adapt to changes to Sphinx going forward. This commit reflects the
behaviour of Sphinx 1.8.4 which is what is packaged into Debian. The
previous config file was generated by a 1.2.x era of Sphinx.

Signed-off-by: Grant Likely <grant.likely@arm.com>