Add new 'chassis-type' root node property
ACPI-based systems generally provide DMI data allowing userspace
software to identify the form-factor of the running system.

Such information can be useful (for example, in order to implement
different behavior on a desktop and on a smartphone) but is not readily
available on DT-based systems.

This patch proposes adding a new optional property to the root node,
allowing a way to specify the system form factor.

The new property is named `chassis-type`, using the same values as
systemd's hostnamed.

Signed-off-by: Arnaud Ferraris <arnaud.ferraris@collabora.com>