Workaround: Delete problematic edge in interrupt tree example
dot has an issue with edges between two "record" shaped nodes that are
directly on top of each other.  This is reported to occur since 2.40
of graphviz.

In v0.3 this device1 -> i-device1 edge was not drawn either but now this
error causes the whole diagram to be missing from the output.

Until dot is fixed or a better work around is found, delete the problematic
edge.  This gets us back to the v0.3 state.

Signed-off-by: Bill Mills <bill.mills@linaro.org>